benchmark/24-DeViEnt/DeViEnt-main/requirements.txt
torch
torchvision
Pillow==10.3.0
transformers==4.41.2
numpy==1.26.4
pandas==2.2.1
tqdm==4.66.4
scikit-learn==1.5.1
wandb==0.17.0